refactor(QueryContext): separate all processing concerns into QueryContextProcessor
